requirements.txt
Django==4.2.13
MarkupPy==1.14
PyJWT==2.8.0
PyYAML==6.0.1
Unidecode==1.3.8
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
diff-match-patch==20230430
django-admin-sortable2==2.2.1
django-auditlog==3.0.0
django-axes==7.0.0
django-ckeditor-5==0.2.13
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.4.2
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-filter==24.2
django-formtools==2.5.1
django-honeypot==1.2.1
django-import-export==4.0.8
django-otp==1.5.4
django-parler==2.3
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-rosetta==0.10.0
django-select2==8.2.1
django-two-factor-auth==1.17.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
idna==3.7
inflection==0.5.1
mysqlclient==2.2.4
openpyxl==3.1.2
packaging==24.1
phonenumbers==8.13.49
pillow==10.3.0
polib==1.2.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
pycparser==2.22
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
rosetta==0.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
xlrd==2.0.1
xlwt==1.3.0